wip, refactor: Use of elution groups as query references and role pushdown
