+++ rust/timsquery/src/models/aggregators/mod.rs
@@ -1,4 +1,7 @@
 pub mod raw_peak_agg;
+pub mod point_agg;
+
+pub use point_agg::RawPeakIntensityAggregator;
 
 pub use raw_peak_agg::{
     MultiCMGStatsAgg,
