+++ rust/timsquery/src/models/base/mod.rs
@@ -1,0 +1,5 @@
+
+pub mod arrays;
+pub mod mz_rt_arrays;
+
+pub use arrays::Array2D;
