+++ rust/timsquery/src/models/mod.rs
@@ -5,5 +5,6 @@ pub mod frames;
 pub mod indices;
 pub mod tolerance;
 pub mod queries;
+pub mod base;
 
 pub use tolerance::Tolerance;
